week3/requirements.txt
groq
pydantic
python-dotenv
pandas
rich
tabulate